fix(cli): validate workspace handle before bp dev deployment
Ensures dev integrations can't be deployed with a workspace handle the
user doesn't own, matching the validation already in place for bp deploy.
