+++ packages/cli/src/command-implementations/deploy-command.ts
@@ -59,7 +59,7 @@ export class DeployCommand extends ProjectCommand<DeployCommandDefinition> {
   }
 
   private async _deployIntegration(api: apiUtils.ApiClient, integrationDef: sdk.IntegrationDefinition) {
-    const res = await this._manageWorkspaceHandle(api, integrationDef)
+    const res = await this.manageWorkspaceHandle(api, integrationDef)
     if (!res) return
     const { integration: updatedIntegrationDef, workspaceId } = res
     integrationDef = updatedIntegrationDef
@@ -519,129 +519,4 @@ export class DeployCommand extends ProjectCommand<DeployCommandDefinition> {
     return fetchedBot
   }
 
-  private async _manageWorkspaceHandle(
-    api: apiUtils.ApiClient,
-    integration: sdk.IntegrationDefinition
-  ): Promise<
-    | {
-        integration: sdk.IntegrationDefinition
-        workspaceId?: string // Set if user opted to deploy on another available workspace
-      }
-    | undefined
-  > {
-    const { name: localName, workspaceHandle: localHandle } = this._parseIntegrationName(integration.name)
-    if (!localHandle && api.isBotpressWorkspace) {
-      this.logger.debug('Botpress workspace detected; workspace handle omitted')
-      return { integration } // botpress has the right to omit workspace handle
-    }
-
-    const { handle: remoteHandle, name: workspaceName } = await api.getWorkspace().catch((thrown) => {
-      throw errors.BotpressCLIError.wrap(thrown, 'Could not fetch workspace')
-    })
-
-    if (localHandle && remoteHandle) {
-      let workspaceId: string | undefined = undefined
-      if (localHandle !== remoteHandle) {
-        const remoteWorkspace = await api.findWorkspaceByHandle(localHandle).catch((thrown) => {
-          throw errors.BotpressCLIError.wrap(thrown, 'Could not list workspaces')
-        })
-        if (!remoteWorkspace) {
-          throw new errors.BotpressCLIError(
-            `The integration handle "${localHandle}" is not associated with any of your workspaces.`
-          )
-        }
-        this.logger.warn(
-          `Your are logged in to workspace "${workspaceName}" but integration handle "${localHandle}" belongs to "${remoteWorkspace.name}".`
-        )
-        const confirmUseAlternateWorkspace = await this.prompt.confirm(
-          'Do you want to deploy integration on this workspace instead?'
-        )
-        if (!confirmUseAlternateWorkspace) {
-          throw new errors.BotpressCLIError(
-            `Cannot deploy integration with handle "${localHandle}" on workspace "${workspaceName}"`
-          )
-        }
-
-        workspaceId = remoteWorkspace.id
-      }
-      return { integration, workspaceId }
-    }
-
-    const workspaceHandleIsMandatoryMsg = 'Cannot deploy integration without workspace handle'
-
-    if (!localHandle && remoteHandle) {
-      const confirmAddHandle = await this.prompt.confirm(
-        `Your current workspace handle is "${remoteHandle}". Do you want to use the name "${remoteHandle}/${localName}"?`
-      )
-      if (!confirmAddHandle) {
-        this.logger.log('Aborted')
-        return
-      }
-      const newName = `${remoteHandle}/${localName}`
-      return { integration: new sdk.IntegrationDefinition({ ...integration, name: newName }) }
-    }
-
-    if (localHandle && !remoteHandle) {
-      const { available } = await api.client.checkHandleAvailability({ handle: localHandle }).catch((thrown) => {
-        throw errors.BotpressCLIError.wrap(thrown, 'Could not check handle availability')
-      })
-
-      if (!available) {
-        throw new errors.BotpressCLIError(`Handle "${localHandle}" is not yours and is not available`)
-      }
-
-      const confirmClaimHandle = await this.prompt.confirm(
-        `Handle "${localHandle}" is available. Do you want to claim it for your workspace ${workspaceName}?`
-      )
-      if (!confirmClaimHandle) {
-        throw new errors.BotpressCLIError(workspaceHandleIsMandatoryMsg)
-      }
-
-      await api.updateWorkspace({ handle: localHandle }).catch((thrown) => {
-        throw errors.BotpressCLIError.wrap(thrown, `Could not claim handle "${localHandle}"`)
-      })
-
-      this.logger.success(`Handle "${localHandle}" is now yours!`)
-      return { integration }
-    }
-
-    this.logger.warn("It seems you don't have a workspace handle yet.")
-    let claimedHandle: string | undefined = undefined
-    do {
-      const prompted = await this.prompt.text('Please enter a workspace handle')
-      if (!prompted) {
-        throw new errors.BotpressCLIError(workspaceHandleIsMandatoryMsg)
-      }
-
-      const { available, suggestions } = await api.client.checkHandleAvailability({ handle: prompted })
-      if (!available) {
-        this.logger.warn(`Handle "${prompted}" is not available. Suggestions: ${suggestions.join(', ')}`)
-        continue
-      }
-
-      claimedHandle = prompted
-      await api.updateWorkspace({ handle: claimedHandle }).catch((thrown) => {
-        throw errors.BotpressCLIError.wrap(thrown, `Could not claim handle "${claimedHandle}"`)
-      })
-    } while (!claimedHandle)
-
-    this.logger.success(`Handle "${claimedHandle}" is yours!`)
-    const newName = `${claimedHandle}/${localName}`
-    return { integration: new sdk.IntegrationDefinition({ ...integration, name: newName }) }
-  }
-
-  private _parseIntegrationName = (integrationName: string): { name: string; workspaceHandle?: string } => {
-    const parts = integrationName.split('/')
-    if (parts.length > 2) {
-      throw new errors.BotpressCLIError(
-        `Invalid integration name "${integrationName}": a single forward slash is allowed`
-      )
-    }
-    if (parts.length === 2) {
-      const [workspaceHandle, name] = parts as [string, string]
-      return { name, workspaceHandle }
-    }
-    const [name] = parts as [string]
-    return { name }
-  }
 }
